Readnow ayyasy

+++ review_buku/models.py
@@ -5,7 +5,7 @@ from book.models import Book
 class Review(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     book = models.ForeignKey(Book, on_delete=models.CASCADE)
-    rating = models.FloatField(default=0)
+    rating = models.IntegerField(default=0)
     comment = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
